Fixed sheet creation, added new API route
Fixed error in sheet creation by template, that would use exercises, tasks & solutions of the sheet used as template instead of creating its own entries.
Fixed display of points for whole exercises in tree visualization.
Added checkbox to select if sheet should be saved as template and therefore not be deleted.